@@ -4,8 +4,11 @@
       <div class="col-md-8">
         <div class="card">
           <div class="card-header">Example Component</div>
-
-          <div class="card-body">I'm an example component!</div>
+          <div class="card-body">
+            <ul>
+              <li>youtube!</li>
+            </ul>
+          </div>
         </div>
       </div>
     </div>
