@@ -8,11 +8,14 @@
                 <div class="card-header">{{ __('Dashboard') }}</div>
                 <div class="card-body">
                     @if (session('status'))
-                        <div class="alert alert-success" role="alert">
-                            {{ session('status') }}
-                        </div>
+                    <div class="alert alert-success" role="alert">
+                        {{ session('status') }}
+                    </div>
                     @endif
                     {{ __('You are logged in!') }}
+                </div>
+                <div id="app">
+                    <example-component></example-component>
                 </div>
             </div>
         </div>
